Refactoring as part of code read
These are minor refactors that I performed while reading the
`NewspaperWorks::Ingest::BatchIngestHelper`.  A major rework was
"privatizing" several methods, which are used only within the `#ingest`
method.  This helps clarify what is the interface for the object.

There are a few outstanding questions regarding using `.create` instead
of `.new`; but those are not things I'm looking to adjust in this read through.