Moving services from app directory to lib
There's no compelling reason for these to be part of the app
directory. And that makes requiring them for isolated tests a bit
trickery.

With this change, we don't need to load the Rails engine to "auto-load"
the underlying services.

Ideally, these would be in the IiifPrint::Services namespace; but that's
perhaps more than we want to do.