Implement IIIF manifest version 2 metadata
This commit will allow both the parent and the child works to be
viewable in the Universal Viewer.  When clicking through the different
images, a user will now see the associated metadata properties that go
with each work.  Currently, each work is sorted by its `Identifier`
though that may become something configurable in the future.  The
metadata properties are exposed and should to be easily configurable by
a developer.  Next steps will be to implement version 3 and also account
for when AllinsonFlex is installed.