Add configuration for sort and refactor some code
Users might not want `Identifier` to always be the sort case so this
commit will expose an easy way to configure what the desired sort by
field will be.  Also, since the IIIF Manifest version is determined
once, we shouldn't need to keep asking what it is.  This refactor will
tell what method to use based what is already known instead of asking
for the version in subsequent steps.