Add base url to `IiifManifestPresenter`
Original implmenetation caused the `IiifManifestPresenter#manifest_for`
method signature to change which would have been an issue if the host
application had an override that also changed its signature.  This
approach extends the `IiifManifestPresenter` to have a base url method
which includs the protocol and hostname which is needed for making the
facet and collection links work when accessed outside of the application
as well as within.