Change model to work
The `model` keyword never sat well with me since in my mind a model
seems like it's referring to the class where as a `work` seems like it's
referring to an instance of the that class.  This commit will rename
occurences of `model` to `work` for semantic reasons.  Also, this commit
will switch the default sorting to `title` instead of `identifier` do to
discussions.