Use a better method of adding an ephemeral class
Some wisdom from Jeremy: "The advantage of this setup is that the class
only lives for the single test. Make sure that around block is scoped to
the tests that need the constant.  What can happen is that you can end
up with multiple declarations of the class (technically re-openings of
the class); which can be fine until one declaration says
`class MyWork < ActiveFedora::Base; end` and then you have class
mismatches.  Alternatively adding that class to the spec_helper.rb makes
it available and only loaded/resolved once.