Fix manifest bug
There is a weird interaction that happens when a child work gets added
to the parent work.  When #get_solr_docs got called, it sometimes
returned 1 solr_doc and when the #find happens, it sometimes returned
nothing which resulted in `nil` being set as the image and passed along.
This would cause issues further down and result in a failed manifest
generation.  I took out @presenter_solr_docs[presenter.id] because it
wasn't being used and seemed to be the source of the problem.