Make UV show up if parent has no file set
This commit will bring over the `WorkShowPresenterDecorator` and
the affected view partial from Louisville.  This decorator allows for
the Universal Viewer to show up on a parent work that has a child work
with a file set, but does not have its own file set.  Previously, the
UV would not render.

Reference file:
https://github.com/scientist-softserv/louisville-hyku/blob/main/app/presenters/hyrax/work_show_presenter_decorator.rb