Removing PDF Composer job and constituent parts
This commit removes portions of the Newspaper Works feature that
stitched together a PDF from provided TIFFs.  An assumption of the IIIF
Print project is that we are not assembling a PDF from the associated
files.  That is the domain of the newspaper_works gem.

Reviewers, please not I *think* we don't need this feature.  And we
should compare this against our proposed feature list (which is
presently being reviewed by other folks).