Moving FileSet and NewspaperIssue definition to specs
Before this commit, the these files were something that all applications
would load into their production code.  Ideally, we don't want to do
that for the two reasons:

- First, not everyone wants a NewspaperIssue model in their application

- Second the file_set, as implemented, directly inherits from
  `ActiveFedora::Base`.  If a downstream FileSet were to not inherit
  from `ActiveFedora:Base` we would have a mismatch error.

With this commit, we're saying that the IIIF print gem will require a
FileSet that conforms to the PCDM expectations.  But we don't dictate
it's implementation details in the downstream application.  We instead
dictate those implementation details in the test application.

Related to:

- https://github.com/scientist-softserv/iiif_print/issues/101