Mocking out polymorphic_url behavior
Prior to this commit (and earlier in development) we had routes and
models for this gem.  However, without custom models and routes, the
polymorphic_url method failed.

We don't need routes, because the purpose of this gem is to provide
configuration options for downstream Hyrax models (e.g. the gem doesn't
ship with it's own Hyrax work types).

With this commit, we mock the behavior of the polymorphic_url and
resolve 3 broken specs.