Tidying up spec_helper and removing spec dependencies
Prior to this commit, we had gems specific for UI testing.  As we've
worked on this gem, we've removed the UI aspects.

The spec times before this change are:

```
Finished in 1 minute 12.18 seconds (files took 7.21 seconds to load)
```

Further, we're not auto-fetching geoname nor lccn nor chronam remote
data for this process.

The spec times after this change are:

```
Finished in 1 minute 7.31 seconds (files took 6.43 seconds to load)
```

It's not much, but in eliminating some of these gems, we improve the
build time as well.

Note, we have commented out lots of specs that were broken.  It is
possible we'll need to reincorporate some of what we removed.

Related to:

- https://github.com/scientist-softserv/iiif_print/issues/43