Removing side-effect causing bug in setting plugins
The primary purpose of this commit is to remediate the following code:

```ruby
def default_plugin_for(file_set)
  if file_set.parent.try(:iiif_print_config?)
    file_set.parent.iiif_print_config.derivative_service_plugins << default_plugin
  else
    [default_plugin]
  end
end
```

As written, the above method was changing the configuration for a work's
parent.  Due to the implementation of the parent's `#iiif_print_config`
it was an isolated update (e.g. updating only the parent object's value
and not the parent object's class's value).

I found this bug by introducting random spec order.

With this commit, I'm doing a few things:

1. Reworking the `plugins_for` method; instead of updating the value, it
   uses that value to set the local `@plugins` instance varaible.
2. Moving methods to group like concepts (e.g. `attr_reader`
   declarations)
3. Extracting `class_attributes` to provide some configurable options.
4. Changing a spec to not test class equality, which does not allow for
   subclass comparison, but to instead test if the object is a
   class (which allows for subclass equality).
5. Leveraging the already existent dependency injection to test some of
   the behavior.
6. Removing the tightly coupled `eq 2` test, which needed to have
   knowledge about the number of `derivative_service_plugins` registered
   for the model.
7. Introducing random rspec run order.  This helps find temporal
   couplings that can cause production issues.