Favoring `class_attribute` over constant
Prior to this commit, we were using a constant and class method to
handle the configurability of the `target_extension`.

With this commit, we remove the constant and provide a class attribute
that allows for configuration and exposes an instance method.

Also, instead of using the terinary args test, I've added a positional
parameter with a default value.