Adjusting method PluggableDerivativeService implementation
Prior to this commit, the `respond_to_missing?` method definition did
not have the correct method signature.  Which could cause failures
elsewhere.

Also, as much as we don't want to repeat knowledge regarding
implementation details, most references to method_missing and the
sibling respond_to_missing duplicate logic; in part to minimize
unnecessary calls.  (See [Always Define respond_to_missing?][1])

I renamed `name` to `method_name` to be consistent with the
parameters of the sibling methods.

And last, I made a change in the logic test to short-circuit faster and
not reference the file_set if we don't have "agreeable" parameters.

References:

- https://github.com/scientist-softserv/iiif_print/issues/43

[1]: https://thoughtbot.com/blog/always-define-respond-to-missing-when-overriding