Favoring instance method for class_attribute
It is almost always better to reduce method chaining (e.g. `a.b.c.d`
moves to `a.d`).  See the [Law of Demeter][1], or the "Principle of
Least Knowledge."

In other words, I can know my neighbor, but I don't know (as well) my
neighbor's neighbor.

This refactor is made possible by introducing a [class_attribute][2] for
`target_extension`.

[1]: https://en.wikipedia.org/wiki/Law_of_Demeter
[2]: https://api.rubyonrails.org/classes/Class.html#method-i-class_attribute