Privatizing methods
Prior to this commit, some of the methods were public when they need
not, nor probably should've been.  (Directly calling method_missing is
an anti-pattern).

With this commit, I'm moving methods to a private state.  Ideally, I'd
like the "services" method to also be private but there's a public test
of that.

Further, I'm ensuring that the passed parameters to the method_missing
are propogated to the allowed_methods.