Favoring instance over class for specs
Prior to this commit, the `FakeDerivativeService` leveraged a class
instance variable.  This is a precarious implementation as we don't
reset the variable in-between runs.

With this commit, we're shifting this "spy" class to be an instance of
the class (thus the tracking values are isolated to each spec run).  We
do this by having the "spy" conform to the interface of a plugin.

Related to:

-  https://github.com/scientist-softserv/iiif_print/issues/43