Reworking include structure to avoid generator
Apologies in advance oh intrepid folk.  This commit contains quite a bit
of conceptual change.

The driving purpose is to look at how to reduce the fragility of our
generator process while also exposing configurable approaches.

First, I am removing generators for each work type and their
corresponding indexer.  This is precarious, in that as folks add new
work types we don't have a mechanism to propagate that.  Second, we were
previously matching files by work_type, which can work, but we can
instead include/prepend logic into those models directly during the
`to_prepare` cycle of the IiifPrint engine.

Second, instead of having a mixin, I'm looking to create configuration
points where you could register a different lineage service.  This helps
create a more crisp separation of concern.

Third, I suspect that some of these "decorations" we do not want to do
unless we configure the model for IiifPrint.  By removing the generator,
we are in greater control of that decision.

All of this because I was trying to figure out how to test the prior
`ancestor_ids` method.

Related to:

- https://github.com/scientist-softserv/iiif_print/issues/43
- https://github.com/scientist-softserv/iiif_print/issues/40