Extracting class attributes to improve configurability
Prior to this commit, we had a duplication of logic regarding the three
types of files we were creating from the alto generation (e.g. xml, txt,
and json).

These are always written and stored as derivatives.  Which is unlikely
to be desirous behavior in all cases.

With this commit, I'm exposing three configuration points with the
primary one being which formats we choose to call and their
corresponding methods on the OCR derivation file.