Moving generated file into engine app
Prior to this commit, we were generating this decorator into the
downstream application.  Thus this engine cedes control to the
downstream application.

However, based on discussions, the `is_child_bsi` may not be in all
implementations.  Which would mean we want this decoration to be
configurable.

With this commit, we favor keeping the decorator inside the engine thus
allowing for future configurability.

Related to:

- https://github.com/scientist-softserv/iiif_print/issues/40