Removing generator injecting code into SolrDocument
Prior to this commit, we injected code into the downstream
`SolrDocument` via the install generator.

With this commit, we eschew adding code via a generator and instead
decoration during the engine `to_prepare` block.

The benefit continues to be that we minimize the surface area of what we
release downstream while we're still working on this gem/engine.

Related to:

- https://github.com/scientist-softserv/iiif_print/issues/121