Favoring configuring Hyrax in to_prepare block
Prior to this commit, we were injecting code into the Hyrax
configuration file of the downstream application.

With this commit, instead of adding code to the file, we're configuring
Hyrax during the engine's to_prepare block.  Further we're adding a
`IiifPrint::Configuration` option for the handling function; to allow
downstream configuration.