Removing install of search builder
Prior to this commit, we do two things:

- Copy downstream a search builder class
- Update the downstream configuration to use that search builder class

With this commit, we no longer copy that search builder class
downstream.  Instead we let the configuration suffice.

This means that downstream implementations can extend (e.g. inherit from
the `IiifPrint::CatalogSearchBuilder`, use it as is, or ignore it).
Regardless, we're not sending that code off on its own to then drift
from our assumptive implementation; thus creating upgrade and
maintenance issues.

Related to:

- https://github.com/scientist-softserv/iiif_print/issues/121