Removing advanced search from IIIF Print
The newspaper_works gem amended the advanced search to include date
range searching.  We brought that along as we cloned the repository to
then begin the gem's adjustments.

As date range searching is not an initial feature of IIIF Print, we can
remove that.  With that it does not make sense for IIIF Print to depend
on blacklight advanced search.

If a downstream implementation wants advanced search, they can look at
how to incorporate that into their application.

Related to:

- https://github.com/scientist-softserv/iiif_print/issues/121