Configuring specs to use file system noid minting
Prior to this commit, we leveraged Hyrax's default noid minting
approach.  It appears that that changed from file based state tracking
to database state tracking.

This created a problem when rebuilding the internal application.  Namely
I encountered the following exception:

```
ActiveRecord::StatementInvalid:
        Could not find table 'minter_states'
```

The option was to either:

1. run the noids-rails generator for the migrations as part of the
   engine cart/install process...
2. set the noids minter class to be the file system.

I have chosen the latter; it's not something I'd want to do for
production but for tests, I'd rather not run another round of
generators.