Removing generated files and commented code
Prior to this commit, we generated code into the downstream application.
This code included two injected modules: `SearchBehavior` and
`AnnotationBehavior`.  In the case of `SearchBehavior` we had switching
logic based on models that came from the `newspaper_works` gem.

With this commit, I removed the `SearchBehavior` installation.  It is not
something we need given the current constraints.

And instead of installing a module to extend the
`BlacklightIiifSearch::Annotation` via the "behavior" pattern, we're
directly prepending a decorator to the class which included the
`AnnotationBehavior` (e.g. cut out the middle module).

Last I removed a commented out line; it was additional chatter we didn't need.

Related to:

- https://github.com/scientist-softserv/iiif_print/issues/121