Adjusting valid plugin testing
Prior to this commit, we were twice checking a file set's parent.  First
in the `Bulkrax::PluggableDerivativeService#plugins_for` methdod and
then immediately in the `Bulkrax::BaseDerivativeService.rb`.

With this commit, we're not repeating that logic.  Instead we're using
the plugins_for method to determine what are the possible valid plugins
for the file_set's parent.  And then we want to again answer what
derivative service is valid for the file set.