Skip blacklight_iiif_search:install solr changes
The [`BlacklightIiifSearch::InstallGenerator`][1] allows for skipping
installing the SOLR changes.  See below:

```ruby
def add_solr_config
  generate 'blacklight_iiif_search:solr' unless options[:'skip-solr']
end
```

With this commit, we'll skip over those changes.  The application will
still have the the suggest routes and configuration; which may need
further investigation: Is the Universal Viewer making requests that
are silently failing?  Do we need the catalog controller configuration?

This all is very challenging to test without a full install.

Related to:

- https://github.com/scientist-softserv/iiif_print/issues/147

[1]: https://github.com/boston-library/blacklight_iiif_search/blob/b2fc5abdb3027762a249dcba9afb8998c6da8ffe/lib/generators/blacklight_iiif_search/install_generator.rb#L43-L45