Narrowing list of default service plugins
Prior to this commit, we were generating several derivatives from the
original file.

In conversations with clients, these extra derivatives are often
confusing and represent additional processing and storage.  The use
cases made more sense for things such as newspapers; in part to
normalize the inputs.

The options are still available, but not as part of the default
configuration.

With this commit, we reduce the default derivatives we run to a small
set.