Refactor Work Show Presenter
The prior implementation processed parent and child works separately to determine while to show the iiif_viewer.
However the child works option was incorrect, as it only
looked for file_set_ids but didn't check file type or
permission.

This refactor fixes that issue, and also splits out a method
which can be overridden to allow for additional file type checks,
and can ultimately be refactored into a config variable.