Remove empty strings from metadata pane
This commit will remove empty strings from showing in the metadata pane.
This occurs in a Bulkrax import when the header exists but there isn't a
value.  Bulkrax would then map that property to `""`.  This would then
show up in the metadata pane as just a label since the value is blank.
Also in the commit, cleaned up some specs and removed `_work` from the
`.default_fields` and `.fields_for_allinson_flex` because it was messing
up the testing.