Introducing derivative locator/applicator concepts
Prior to this commit, I was passing an array of strings back from the
locator.  With that pathway, I was ignoring how the existing
`Hyrax::FileSetDerivativesService` would likely be used as a fallback.

With this commit, I'm beginning to account for how that interaction
would be; in part by focusing on a known use case and building the
interactions accordingly.

This commit stencils in the expected behavior in what I think is
adequate to convey intention and design.

I also begin demonstrating how the original configuration can be
leveraged to do the derivative work.  This still assumes a configuration
for derivatives; something that I continue to punt on until I have more
concrete code that I can further build from.

Related to:

- https://github.com/samvera-labs/bulkrax/issues/760
- https://github.com/samvera-labs/bulkrax/pull/761
- https://github.com/scientist-softserv/utk-hyku/issues/343