[i335] - port over serverless code from nnp (#198)
* [i335] - port over serverless code from nnp

* [i335] - implement serverless iiif, move code into  DisplayImagePresenterBehavior

TODO: Loading issue in dev; when we change code the app no longer hits our iiif_manifest_presenter_behavior.rb and reverts back to using the app instead of cloudfront. We also added the prepend in the after_intitialize because the app couldn't find the digest method otherwise.

* [i335] rubocop fix

* Add test and refactor

This commit will add a test for #display_image.  Also, a refactor was
introduced to account for both Hyrax 2 and 3.  The method signature
of the `Hyrax.config.iiif_image_url_builder` lambda changed to add an
additional argument in Hyrax 3.

ref: https://github.com/samvera/hyrax/blob/2.x-stable/lib/hyrax/configuration.rb#L397-L399
ref: https://github.com/samvera/hyrax/blob/3.x-stable/lib/hyrax/configuration.rb#L250-L252

* Refactor `DisplayImagePresenterBehavior`

This commit will add OVERRIDE comments where useful and add a blurb in
the README for this new feature.  The #display_content method was added
to support presentation 3 manifests.  `SERVERLESS` has been renamed to
`EXTERNAL`.  `DisplayImagePresenterBehavior` has beeen refactored to
incorporate existing code.  Specs have also been refactored.

* add EXTERNAL_IIIF_URL instructions to the README

---------

Co-authored-by: Kirk Wang <kirk.wang@scientist.com>
Co-authored-by: Rob Kaufman <rob@notch8.com>