Optimize identify -verbose
The `identify -verbose` was too expensive of a method because it is
getting back a ton of information that wasn't even being used.  We are
opting for `identify -format` to pull only the information we are
looking to use instead to help speed up the `CreateDerivativesJob`
process.

The output looks something like:

```sh
Geometry: 5100x6592
Depth: 8
Colorspace: sRGB
Alpha: Undefined
MIME Type: JPEG
```