Extracting ancestry identifier configurability
Prior to this commit, we assumed that a work's `id` was the mechanism
for always identifying a work in regards to it's ancestry/lineage.

However, for implmentations that incorproate slug behavior, that reality
is not always the case.

With this commit, we expose a configurable mechanism for altering the
ancestry functionality.

Related to:

- https://github.com/scientist-softserv/adventist-dl/pull/354/
- https://github.com/scientist-softserv/adventist-dl/issues/319