Add a highlight detection
The `#has_any_highlighting?` method is used to determine if the search
returned any highlighting. If it did, then we add the parent_query to
the search params so that the search results.  This will avoid causing
and No matches found notification in the UV search.  Also in this
commit, we change `hl.fl` to return all fields that are indexed.