Add IIIF Helpers into the load path
This commit makes the IIIF Helpers work by loading in the engine.rb.
The main reason for the change is the get the #uv_search_param method
usable for the Universal Viewer.  This parameter will automatically get
the user into the search term they were using for the catalog search.