Add configuration for OCR coordinate search
This commit adds a configuration option for the OCR coordinate search
which will allow applications to specify where to pull the JSON
coordinate data.  By default it pulls it from the json file that is
created through the TextExtractionService.  However, if the application
holds the coordinate data in a different location, such as on a Solr
field, then that can now be configured.

In Essi, they do not want any derivative service so the .json, .xml, and
.txt files would not be created thus by default, the OCR cannot be
searched.  They are indexing the coordinates in a Solr field on the Solr
docment instead of reading from file every time.