Add configuration for all_text indexing
This commit adds a configuration for the all_text_* fields.  In Essi,
they do not use all_text_tsimv nor all_text_timv so the full text
catalog search would not work without it.  The default way the full text
is being captured in IIIF Print is by reading the txt file that is
generated through the TextExtractionDerivativeService.  Essi does not
use that service but instead has their own implementation of generating
the alto xml.  We leverage that implementation here and use a lambda to
to extract full text from the alto xml.