Add a way to sort Allinson Flex metadata
This configuration allows you to sort the metadata within an application
that uses Allinson Flex.  In a non-Allinson Flex application, you would
use the IiifPrint.config.metadata_fields to sort the metadata.  However,
in Allinson Flex, the fields come in according to the m3 profile.  This
configuration allows you to use an Array of Symbols to sort the metadata
in an intentional way.

Some profiles have fields with the same name repeated.  This commit will
remove repeated fields though this functionality just takes the first
field and ignores the rest.  This isn't very controlled at the moment
but works for Essi's situation.

Ex. Imagine if you had fields of:
[
  #<AllinsonFlex::ProfileProperty id: nil, name: "title", indexing: ["stored_searchable"]>,
  #<AllinsonFlex::ProfileProperty id: nil, name: "title", indexing: ["stored_searchable"]>
 ]

The first object would yield
Title: My Sweet Title

and the second field would yield
Photograph Information: My Sweet Title

The second field would be ignored since the value is the same as the
first field.  This is how Essi displays their metadata so we will use
this method until told otherwise.