Add a return if parent and child metadata match
Prior to this commit, the metadata for the parent image would display
both on the manifest and canvas levels. In this commit, we check if the
image we are looking at is on the parent or the child.  If it is on the
parent, we won't display it on the canvas level because it is already
being displayed on the manifest level.