🐛 Fix additional query terms removal
This commit will fix the removal of the additional query terms.  Here we
use a regular expression to remove the additional query terms from the
query string instead of using the `gsub` method.  This is more accurate
because it was noticed that `document['is_page_of_ssim']` isn't always
what we expect.  It does not work in a nested work context.