⚙️ Adding derivative_rodeo as dev dependency
Why as a development dependency?  Because the DerivativeRodeo introduces
a dependency on Faraday >= 1.  And the Valkyrie and ActiveFedora
versions which Hyrax 2 and 3 depend on have a Faraday dependency of < 1.

I am pushing this up so that I can begin development on the ingest
aspect of the Derivative Rodeo.  Also to see how this resolves in our CI
setup and to see the impact, if any on downstream implementations of
IIIF Print (e.g. Adventist, British Library, ATLA, PALNI/PALCI, UTK, and
others).

The plan is to determine if we want to have this Faraday conflict setup
or if we want to swap out something else in the underlying
DerivativeRodeo.

Related to:

- https://github.com/scientist-softserv/adventist-dl/issues/330
- https://github.com/scientist-softserv/iiif_print/issues/219
- https://github.com/scientist-softserv/iiif_print/issues/220