♻️ Moving PDF file split guards into the service
As I work on the Derivative Rodeo handling of PDFs, consolidating the
logic to be within the refactored service.  This allows for more easily
checking the Derivative Rodeo for a corresponding PDF that has already
been pre-processed.

This is a significant refactor in that I want a keen eye reviewing the
changes to ensure I haven't introduced a logic bug in the
transformation.

Related to:

- https://github.com/scientist-softserv/iiif_print/issues/219p