♻️ Preparing Splitter Service Swap for DerivativeRodeo
This commit is a refactor in-place.  The primary goal is to allow for
passing the file_set to the child works; something that is ideal for the
derivative rodeo's interface.

This is intended to be a swap-in-place change.  That is to say, if we
deploy this change and have already enqueued jobs, nothing will fail nor
break.  The past enqueued jobs (with a work) will use the work based
logic but future enqueueings will use file_set.

In using the file_set, we also avoid the issue of having passing a `nil`
as the parent, and thus creating an infinite rescheduling cycle.

Related to:

- https://github.com/scientist-softserv/iiif_print/issues/220