🐛 Correct search count
A funky bug was observed when on a parent work of a PDF, the count was
off by one.  It registers OCR hits as both an OCR hit and a metadata
hit.  This is likely because of adding snippets in the search since all
the file sets' texts need to be indexed on the parent work as well.  The
parent work essentially would double all the texts found.  This commit
is a bit hacky but it removes that extra hit while keeping functionality
for both OCR hits and metadata hits.  This is a  bit of future proofing
since it only would happen in applications with snippets enabled.