🐛 Fix bug OCR search bug with AllinsonFlex
This commit fixes a bug where the OCR search was not working with
AllinsonFlex.  The problem was the `qf` parameter was getting extra
fields added but was missing a `' '` between the fields default fields
and the AllinsonFlex fields.  Also in this fix, I realize that I was
trying to decorate the `IiifSearchBuilder` when it is actually a file
that the generator installs.  Since the file is in Blacklight, I opted
to gsub the content instead.