🐛 Adding more robust (albeit hacky) handling of ancestors
This commit includes three substantive changes:

- Extracting a method for determining the directory in which to look for
  a file.
- Addressing a very odd bug found in `public_send` behavior.
- Adding a fallback when we don't have an ancestor.

First, the fallback.  It is hacky (as noted by the `HACK:` comment).
But it's quick and for Adventist is suitable.  We could extract a
function that is the fallback and then call that.

Second, the odd bug manifested as `nil` being raised as an exception;
that is to say with the original `rescue => e` the `e` was `nil` (and
it's class was `NilClass`).

I found that if I removed the
`IiifPrint.public_set(ancestor_method, file_set)` and made an explicit
call, then the exception went away; albeit now returning `nil`.

Which leads to the `HACK:` for handling `nil`.

Last, in extracting the method, I leave a public place for more specific
handling of the possible directory for the preprocessed derivative files.

Related to:

- https://github.com/scientist-softserv/derivative_rodeo/issues/56