🎁 Add graceful fallback of preprocessing
Prior to this commit, if we did not actually have preprocessed file in
the remote system, we'd encounter errors with files missing; we assumed
we could "copy" the preprocessed url to the local file.

With this commit, we check to see if the preprocessed file exists.  Then
fallback to a specified FileSet#remote_url and the fallback even
further.

The specs define the expected behavior.

Closes #282

Related to:

- https://github.com/scientist-softserv/iiif_print/issues/282