🐛 Leverage CopyGenerator to ensure we have local copy
We need to address the case where SpaceStone failed to copy the file
into S3.  When the file does not exist in S3, we can check if we have
the file locally and use that as the pre-processed location.  Otherwise,
we can copy the original remote file and proceed.

Related to:

- https://github.com/scientist-softserv/iiif_print/pull/283
- https://github.com/scientist-softserv/iiif_print/issues/282