🧹 Add Solargraph as dependency
I have used the following to get jump to definition and render
documentation for the symbol at cursor.

1.  Add solargraph as a development dependency
2.  Run `bundle`
3.  Run `bundle exec yard gems` (or target specific gems)
4.  Add "relevant" gems to the `.solargraph.yml`

The above should *not* be in the docker container.  You would run this
in the environment where your editor is running.  There is probably a
way to make that work with Docker, but I’m not here for that today.