Service to split previously ingested pdf
This creates a new service which combines:
- IiifPrint::SplitPdfs::DestroyPdfChildWorksService
- IiifPring::Jobs::ChildWorksFromPdfJob

Beginning with a PDF fileset, it removes any existing child works (found
by either the fileset ID or the title), any pending
relationship table entries (found by parent id and fileset ID), and then
submits a new job to do the pdf splitting.

Note that pending relationship entries will not be removed if they don't
have the file_id (which is the id of the fileset that spawned them). This
is to avoid removing pending relationships that could still be needed for
another fileset on the parent work.