🧹 Add .internal_test_app
Prior to this commit, when we'd push this to CI, we'd continue to
rebuild the `.internal_test_app`; this resulted in 2 hour build times.

By adding it to the repository, we explicitly say when we want to
rebuild the internal test application (e.g. by deleting it and letting a
local `bundle exec rake` rebuild it).

See the README.md for more details.