🐛 Ease how we determine preprocessed location
Prior to this commit, we assumed the ancestor must have an AARK_ID.
However that is not always the case.  Which is confounding given the
data structure of the files; however such is the way of the world.

With this commit we fallback to the file_set's internal information to
attempt to find the file in the preprocessed location.  Namely if the
FileSet had an import_url, we'll use that to derive where it probably
went in SpaceStone.

Note: sniffing out where this file exists in SpaceStone is a very
fragile process.

Closes #289

Related to:

- https://github.com/scientist-softserv/iiif_print/issues/289