🐛 Work through fallback of rodeo URLs
This commit replaces raising exceptions when we can't derive the URL.
Instead it opts to return nil and let the caller determine what to do.

The primary result is that it's easier to have a rodeo configuration
"miss" finding files in the pre-process location due to configuration
issues.

But alas, that's the nature of a rodeo.

Related to:

- https://github.com/scientist-softserv/iiif_print/issues/294